Rework the "database-less" tests.
It's not possible for DbLinq_test.dll to test ~all databases, as the
ObjectFactory Reflection-based Provider lookup mechanism is not robust --
there needs to be a concrete reference to the assembly providing the provider
in order for the provider to be found.  Since e.g. DbLinq.SqlServer.dll has a
provider, this would require that DbLinq_test.dll have an assembly reference
to DbLinq.SqlServer.dll, which is BAD for small, targeted tests.

Consequently, the provider-specific tests will need to include the
database-less tests (or use new projects, as was done with the
DbLinq.SqlServer_test_ndb project).

	* src/DbLinq.SqlServer/Test/DbLinq.SqlServer_test_ndb.csproj: Added;
	  database-less NUnit tests for DbLinq.SqlServer.dll.
	* src/DbLinq.SqlServer/Test/MsSqlDataContractTest.cs: Moved from
	  src/DbLinq/Test; add code to ensure we get an assembly reference to
	  DbLinq.SqlServer.dll.
	* src/DbLinq/Test/DataContextTest.cs: 
	  - Move tables (Person, BadPerson) into PeopleTable.cs
	  - Move CreateDataContext() logic into 
	    src/DbLinq/Test/Providers/DataContextTestBase.cs.  The tests in
	    this file must be provider-agnostic, e.g. expected exceptions for
	    invalid arguments, etc.
	* src/DbLinq/Test/DbLinq_test.csproj: Deal with file movement, etc.
	* src/DbLinq/Test/PeopleTable.cs: Added; tables for use by
	  DataContextTestBase & related tests.
	* src/DbLinq/Test/Providers/DataContextTestBase.cs: Added; base class
	  for provider-specific database-independent tests.
	* src/DbLinq_StudioExpress.sln: Flush.