Fix ReadTest.C22_SelectEmployee_GetCommandTextWithNoFilter().
The problem is that if you do:

	db.GetCommand(db.GetTable<Employee>()).CommandText

you haven't added any expressions to the IQueryable<Employee>, so
QueryProvider.ExpressionChain.Expressions.Count is 0, which causes
ArgumentOutOfBoundsException from QueryBuilder.BuildExpressionQuery().

The fix we're employing here is that if there are no expressions on the
IQueryable<T>, then we'll add an "identity" expression: an expression that
just returns the input expression unchanged.

This also adds a ReadTest.C23_SelectEmployee() test which currently fails (due
to Binary throwing an argument null exception -- looks like Binary needs some
special handling, as Linq 2 SQL doesn't instantiate the Binary object if the
column value is null).