Fix DynamicLinqTest tests.
The tests are correct; DbLinq is broken.

For DyanmicLinqTest.DL9_PredicateBuilderCount() and
DynamicLinqTest.DL10_PredicateBuilderWhere(), queries similar to the following
are created:

	db.Customers.Where(c => true && c.City == "Paris")

The problem is the "extra" 'true' was causing SQL like the following to be
generated:

	SELECT * FROM Customers WHERE 1 AND City = 'Paris'

"1 AND" isn't liked by SQLite and SQL Server, hence the failure.

We fix this within ExpressionOptimizer.AnalyzeAndAlso() by looking for the 
binary expressions "true && rhs" and "lhs && true", and replacing the binary 
expression with just rhs/lhs (as appropriate), thus removing the boolean 
constant expression.