"Fix" ReadTest_Complex test for SQLite.
The problem is that the SQLite Northwind DB is incomplete (compared to the SQL
Server version), and the 
ReadTest_Complex.D2_SelectPensByLocalPropertyAndConstant() test is checking
for a 'Chai' product, which the SQLite DB doesn't have.

The fix is to choose a product that *does* exist in both the SQLite and SQL
Server Northwind DBs, in this case 'Carnarvon Tigers'.

Rename the test to
ReadTest_Complex.D2_SelectProductByLocalPropertyAndConstant() since we're no
longer selecting a pen.