Provide a *real* Northwind database for SQLite.
Several of the unit tests were failing under SQLite because they were 
asserting data that didn't exist in the SQLite version, but did exist in 
SQL Server, e.g. Products with a ProductName of 'Chai' (see r1132).

This was getting annoying enough that I "ported" the original Northwind 
data to SQLite, so that at least SQLite and SQL Server can use the same 
data for tests.

	* LICENSE: Added; file describing the various licenses used.
	* LICENSE.MSPL: Added; Ms-Pl text.
	* LICENSE.X11: Added; MIT/X11 license text.
	* tests/instnwnd.SQLite.sql: Added; creates the Northwind data 
	  set for SQLite.
	* tests/instnwnd.SqlServer.sql: Added; Ms-Pl licensed Northwind 
	  data set; source file for instnwnd.SQLite.sql.
	* tests/Northwind.db3: SQLite3 Northwind database file.
	* src/DbLinq.Sqlite/Test/TestBase.cs: SQLite uses Ordinal string
	  comparisons, not InvariantCulture comparisons (fixes
	  ReadTest.D04_SelectProducts_OrderByName).
	* examples/DbLinq.SQLite.Example/nwind/Northwind.cs: Northwind 
	  uses a 'Region' table, not a 'Regions' table.  Fix.