Add partial support for method calls within selects.
It's possible for a method call to be within a select:

	from e in db.Employees select new {
		A = DoSomething(e),
		/* ... */
	};

This is all well and good, but it would result in an ArgumentException saying
that the method DoSomething() was not supported.  Oops.

Fix this, but the support is incomplete: only static methods are currently
supported.  (In ReadTest.C26_SelectWithNestedMethodCall(), note that
GetInstanceName() and e.EmployeeTerritories.ToList() are commented out, as
both of these are currently unsupported.)

We fix this by replacing a ParameterExpression with the TableExpression that
is the type of the ParameterExpression.  This results in the expression tree
going from 'A = DoSomething(e)' to 'A = DoSomething(new Employee {...})'.

Instance methods (and likely full closures) aren't supported because they use
parameters which don't exist within the DbLinq-generated expression tree.