Fix boolean parameters.
Mono.Options.OptionSet is apparently schizo when it comes to boolean
parameters; these are NOT the same:

	{ "foo", (bool v) => save = v }
	{ "bar", v => save = v != null }

The former is what was used, the latter is the normal boolean option
mechanism.

The problem is that (bool v) requires that 1. the option have a value, and 
2. the value be (somehow) convertible to bool.

However, due to "historical" reasons (boolean parameters were supported before
built-in type conversion support), boolean options would get either null or
non-null values (and the non-null value would be the option name).  Neither
'null' nor the non-null string would be convertible to bool, as (2) requires.

The end result: 'DbMetal.exe /pluralize ...' results in an exception.  Ouch.

Following conventional bool option style allows -pluralize/etc. to work.