Support nullable types in Bulk Inserts for SQL Server.
Patch idea thanks to yuriastrakhan.

Fixes issue 124.

Issue 124 also mentions an issue with MySQL.  I wonder if this is still the
case, as MySqlVendor.BulkInsert<T>() uses
IDbDataParameterExtensions.SetValue() to set the IDbDataParameter.Value
property, and will set it to DBNull.Value if the value is null.  In theory,
this should allow nullable types to be properly supported by MySQL.