Fix invalid code generated by `DbMetal --generate-equals-hash`.
The precedence of C# exclusive or (^) doesn't allow easy chaining; expressions
containing exclusive-or need to be wrapped in parenthesis for proper parsing.

This also changes the behavior of Equals() to use
EqualityComparer<T>.Default.Equals(a, b) instead of object.Equals(a, b), as
the former will properly deal with types that implement IEquatable<T> (and
will fall back to object.Equals() if necessary).  It also avoids boxing.

Patch thanks to Chip <chipboling@gmail.com>.

Fixes issue 220.