Allow port names to be specified within `DbMetal /server`.
For example:

	DbMetal /server:hostname:port ...

will connect to server "hostname" on port "port".

Previously this would fail because most databases want the port number in a
distinct Port key/value pair in the connection string.  This is now done
(except for SQL Server, which uses a comma to separate the hostname from the
port number).

Fixes Issue 184.