CodeDom: Ensure that entity fields are always prefixed with '_'.
The problem is this:

	DbMetal /provider:Sqlite /language:VB /code:nwind.vb \
		src/DbLinq.MySql\Test\NorthwindMySql.dbml
	vbc /t:library nwind.vb /d:MONO_STRICT /r:System.Data.dll \
		/r:System.Data.Linq.dll

VB.NET is a case-insensitive language, and NorthwindMySql.dbml uses case to
differentiate between Storage and Member values.  Result: *lots* of
compilation errors from `vbc` as it attempts to bind to the (private) fields
instead of the (public) properties of the "same" name (e.g. e.regionId vs.
e.RegionId).

Ensuring that field names are prefixed with '_' enables VB.NET support.